:recycle: Merge ProcessModelConverter into ProcessModelService
