+++ src/converters/index.ts
@@ -1,5 +1,4 @@
 export * from './empty_activity_converter';
 export * from './event_converter';
-export * from './process_model_converter';
 export * from './user_task_converter';
 export * from './manual_task_converter';
